firebase/requirements.txt
pyserial
flutter_admin
dotenv
